나무 자르기

+++ 백준/2805.cpp
@@ -1,9 +1,11 @@
 #include <iostream>
 #include <vector>
 using namespace std;
-// 필요로한 나무를 얻기위한 최소한의 나무길이를 구해라. 
 
-long long BinarySearch(vector<long long> &arr, long long &N, const long long &begin, const long long &end)
+// 설명 blog
+// ggh-png.github.io 
+
+int BinarySearch(vector<long long> &arr, int &N, const long long &begin, const long long &end)
 {
     long long m = (begin + end) / 2;
     if(begin == end)
@@ -23,7 +25,7 @@ int main()
     ios_base :: sync_with_stdio(false); 
     cin.tie(NULL); 
     cout.tie(NULL);
-    long long n, m;
+    int n, m;
     cin >> n >> m;
     vector<long long> arr(n);
     // 나무 입력 받기 
@@ -34,11 +36,8 @@ int main()
         cin >> arr[i];
         if(arr[i] > max)
             max = arr[i];
-
     }
     min = max - m;
-    // begin : 0 end :n max
     cout << BinarySearch(arr, m, min, max) << endl;
- 
     return 0;
 }
